Promote slo commands from alpha to main
