+++ cmd/reliably/slo/cmd.go
@@ -2,8 +2,10 @@ package slo
 
 import (
 	"github.com/MakeNowJust/heredoc/v2"
+	"github.com/reliablyhq/cli/cmd/reliably/slo/agent"
 	init_cmd "github.com/reliablyhq/cli/cmd/reliably/slo/init"
 	"github.com/reliablyhq/cli/cmd/reliably/slo/report"
+	"github.com/reliablyhq/cli/cmd/reliably/slo/sync"
 	"github.com/spf13/cobra"
 )
 
@@ -14,8 +16,10 @@ func NewCommand() *cobra.Command {
 		Long:  longDescription(),
 	}
 
+	cmd.AddCommand(agent.NewCommand(nil))
 	cmd.AddCommand(init_cmd.NewCommand(nil))
 	cmd.AddCommand(report.NewCommand(nil))
+	cmd.AddCommand(sync.NewCommand(nil))
 	return &cmd
 }
